Separando os callbaks em arquivos
