+++ src/controller/controller.c
@@ -82,37 +82,6 @@ void execute( GtkWidget *widget, gpointer data ) {
 
 // CALLBACKs of video_page
 
-void set_repository( GtkWidget *widget, gpointer data ) {
-    gource_settings.video.repository = gtk_file_chooser_get_filename( GTK_FILE_CHOOSER( widget ) );
-}
-
-gboolean set_title( GtkWidget *widget, gpointer data ) {
-    if( !string_copy_char_array( gource_settings.video.title,
-                                 gtk_entry_get_text( GTK_ENTRY( widget ) ) ) ) {
-        free_memory( &gource_settings, true );
-    }
-    return FALSE;
-}
-
-void set_screen_mode( GtkWidget *widget, gpointer data ) {
-    gource_settings.video.screen_mode =
-        gtk_combo_box_text_get_active_text( GTK_COMBO_BOX_TEXT( widget ) );
-}
-
-void set_background_color( GtkWidget *widget, gpointer data ) {
-    GdkRGBA rgba;
-    gtk_color_chooser_get_rgba( GTK_COLOR_CHOOSER( widget ), &rgba );
-    if( !string_copy_char_array( gource_settings.video.background_color,
-                                 rgba_to_hex( rgba.red, rgba.green, rgba.blue ) ) ) {
-        free_memory( &gource_settings, true );
-    }
-}
-
-void set_camera_mode( GtkWidget *widget, gpointer data ) {
-    gource_settings.video.camera_mode =
-        gtk_combo_box_text_get_active_text( GTK_COMBO_BOX_TEXT( widget ) );
-}
-
 // CALLBACKs of subtitle_page
 
 void set_subtitle_file( GtkWidget *widget, gpointer data ) {
@@ -215,7 +184,6 @@ void string_tolower( char *string ) {
 }
 
 void add_to_argv_valid_field( _gource *gource, char **argv, int *size ) {
-    char *tmp;
     prepare_color( gource->video.background_color );
     prepare_color( gource->subtitle.color );
     if( gource->video.repository != NULL ) {
